+++ packages/expo-router/src/imperative-api.tsx
@@ -1,5 +1,6 @@
 import type { ImperativeRouter } from './global-state/router';
-import { router } from './global-state/router';
+import { router as internalRouter } from './global-state/router';
 
 export type { ImperativeRouter };
-export { router };
+// Hide internal `goBack` and `linkTo` methods from the public API and typedoc.
+export const router: ImperativeRouter = internalRouter;